Added more contries and improved UX
Firstly, if the user is stuck, they can type ? to get one letter unlocked. This will cost them 3 guesses rather than 1 per guess. In addition a guess is added for every guess even if the letter is in the country. The guess will not be added if the country is unlocked.